Changed sandbox terrain, implemented a new system, code fix
Changed the sand box terrain system so that it will give more authentic feel near to the final product. Implemented weapon zoom system so that you can "aim down" and fixed the enemy hit on weapon script. It will specifically damage the hit target instead of the script relying on the hierarchy